fix(forge): format `new` expressions (#6408)
* fix(forge): format `new` expressions

* fmt

* style: format

* fix(forge): use write_chunk for `Expression::New`

* style: format Prank.t.sol with new formatter

* style: format formatter.rs

* no trailing semicolons

---------

+++ crates/fmt/testdata/ArrayExpressions/fmt.sol
@@ -4,7 +4,7 @@ contract ArrayExpressions {
         uint256[10] memory sample;
 
         uint256 length = 10;
-        uint256[] memory sample2 = new uint[](length);
+        uint256[] memory sample2 = new uint256[](length);
 
         uint256[] /* comment1 */ memory /* comment2 */ sample3; // comment3
